+++ src/action.rs
@@ -36,7 +36,7 @@ pub enum Action {
     //            |          | the bare content for a message to be formed
     //            | the destination authority
     // RoutingNode will form the RoutingMessage and sign it as its own identity
-    Churn(::direct_messages::Churn, Vec<::crust::Endpoint>),
+    Churn(::direct_messages::Churn, Vec<::crust::Endpoint>, ::NameType),
     SetCacheOptions(CacheOptions),
     Terminate,
 }