+++ src/routing_node.rs
@@ -102,14 +102,15 @@ impl RoutingNode {
             client_restriction: client_restriction,
             action_sender: action_sender.clone(),
             action_receiver: action_receiver,
-            event_sender: event_sender,
+            event_sender: event_sender.clone(),
             filter: ::filter::Filter::with_expiry_duration(Duration::minutes(20)),
             connection_filter: ::message_filter::MessageFilter::with_expiry_duration(
                 ::time::Duration::minutes(20)),
             core: core,
             public_id_cache: LruCache::with_expiry_duration(Duration::minutes(10)),
-            accumulator: MessageAccumulator::new(),
-            refresh_accumulator: RefreshAccumulator::new(),
+            accumulator: MessageAccumulator::with_expiry_duration(::time::Duration::minutes(5)),
+            refresh_accumulator: RefreshAccumulator::with_expiry_duration(
+                ::time::Duration::minutes(5), event_sender),
             cache_options: CacheOptions::no_caching(),
             data_cache: None,
         }
@@ -128,10 +129,11 @@ impl RoutingNode {
                     let _ = self.send_content(our_authority, to_authority, content);
                 },
                 Ok(Action::ClientSendContent(to_authority, content)) => {
+                    debug!("ClientSendContent received for {:?}", content);
                     let _ = self.client_send_content(to_authority, content);
                 },
-                Ok(Action::Churn(our_close_group, targets)) => {
-                    let _ = self.generate_churn(our_close_group, targets);
+                Ok(Action::Churn(our_close_group, targets, cause)) => {
+                    let _ = self.generate_churn(our_close_group, targets, cause);
                 },
                 Ok(Action::SetCacheOptions(cache_options)) => {
                     self.set_cache_options(cache_options);
@@ -400,7 +402,7 @@ impl RoutingNode {
         }
 
         let message = signed_message.get_routing_message().clone();
-        let mut message_digest = ::filter::Filter::message_digest(&message);
+        let message_digest = ::filter::Filter::message_digest(&message);
 
         // Cache a response if from a GetRequest and caching is enabled for the Data type.
         self.handle_cache_put(&message);
@@ -473,7 +475,7 @@ impl RoutingNode {
                             None => return Err(RoutingError::UnknownMessageType),
                         }
                     }
-                    InternalRequest::Refresh(type_tag, bytes) => {
+                    InternalRequest::Refresh(type_tag, bytes, cause) => {
                         let refresh_authority = match our_authority {
                             Some(authority) => {
                                 if !authority.is_group() { return Err(RoutingError::BadAuthority) };
@@ -481,14 +483,11 @@ impl RoutingNode {
                             },
                             None => return Err(RoutingError::BadAuthority),
                         };
-                        // FIXME (ben 8/09/2015) Exclude refresh message from being blocked
-                        // after succesful resolution
-                        message_digest = None;
                         match *signed_message.claimant() {
                             // TODO (ben 23/08/2015) later consider whether we need to restrict it
                             // to only from nodes within our close group
                             Address::Node(name) => self.handle_refresh(type_tag, name, bytes,
-                                refresh_authority),
+                                refresh_authority, cause),
                             Address::Client(_) => Err(RoutingError::BadAuthority),
                         }
                     }
@@ -563,8 +562,7 @@ impl RoutingNode {
         let skip_accumulator = match message.content {
             Content::InternalRequest(ref request) => {
                 match *request {
-                    InternalRequest::Refresh(_, _) => {
-                        println!("SKIPPED ACCUMULATOR FOR REFRESH");
+                    InternalRequest::Refresh(_, _, _) => {
                         true
                     },
                     _ => false,
@@ -625,10 +623,11 @@ impl RoutingNode {
 
     // ---- Churn ---------------------------------------------------------------------------------
 
-    fn generate_churn(&self, churn: ::direct_messages::Churn, target: Vec<::crust::Endpoint>)
-        -> RoutingResult {
+    fn generate_churn(&mut self, churn: ::direct_messages::Churn, target: Vec<::crust::Endpoint>,
+        cause: ::NameType) -> RoutingResult {
         debug!("CHURN: sending {:?} names to {:?} close nodes",
             churn.close_group.len(), target.len());
+        self.refresh_accumulator.register_cause(&cause);
         // send Churn to all our close group nodes
         let direct_message = match ::direct_messages::DirectMessage::new(
             ::direct_messages::Content::Churn(churn.clone()),
@@ -641,9 +640,7 @@ impl RoutingNode {
             ignore(self.connection_manager.send(endpoint, bytes.clone()));
         }
         // notify the user
-        if self.core.routing_table_size() >= ::types::QUORUM_SIZE {
-            let _ = self.event_sender.send(::event::Event::Churn(churn.close_group));
-        };
+        let _ = self.event_sender.send(::event::Event::Churn(churn.close_group, cause));
         Ok(())
     }
 
@@ -964,6 +961,7 @@ impl RoutingNode {
     // ----- Send Functions -----------------------------------------------------------------------
 
     fn send_to_user(&self, event: Event) {
+        debug!("Send to user event {:?}", event);
         if self.event_sender.send(event).is_err() {
             error!("Channel to user is broken. Terminating.");
             let _ = self.action_sender.send(Action::Terminate);
@@ -1133,11 +1131,12 @@ impl RoutingNode {
     fn handle_refresh(&mut self, type_tag: u64,
                                  sender: NameType,
                                  payload: Bytes,
-                                 our_authority: Authority) -> RoutingResult {
+                                 our_authority: Authority,
+                                 cause: ::NameType) -> RoutingResult {
         debug_assert!(our_authority.is_group());
         let threshold = self.group_threshold();
         match self.refresh_accumulator.add_message(threshold,
-            type_tag.clone(), sender, our_authority.clone(), payload) {
+            type_tag.clone(), sender, our_authority.clone(), payload, cause) {
             Some(vec_of_bytes) => {
                 let _ = self.event_sender.send(Event::Refresh(type_tag, our_authority,
                     vec_of_bytes));