+++ CHANGELOG.md
@@ -1,4 +1,18 @@
 # Routing - Change Log
+
+## [0.3.11]
+- [#696](https://github.com/maidsafe/routing/pull/696) expose NAME_TYPE_LEN and random traits
+- [#695](https://github.com/maidsafe/routing/pull/695) correct style error in error.rs
+- [#692](https://github.com/maidsafe/routing/pull/692) add cause and event::DoRefresh for improvements to churn
+- [#691](https://github.com/maidsafe/routing/pull/691) update QA libsodium documentation
+- [#690](https://github.com/maidsafe/routing/pull/690) correct failing test
+- [MAID-1361](https://maidsafe.atlassian.net/browse/MAID-1361) unit tests for id, public_id, error, data, direct_messages
+- [MAID-1356](https://maidsafe.atlassian.net/browse/MAID-1356) unit test filter.rs
+- [MAID-1358](https://maidsafe.atlassian.net/browse/MAID-1358) unit test signed_message
+
+## [0.3.10]
+- [#685](https://github.com/maidsafe/routing/pull/685) use latest accumulator
+
 ## [0.3.9]
 - [MAID-1349](https://maidsafe.atlassian.net/browse/MAID-1349) refresh_request to use authority
 - [MAID-1363](https://maidsafe.atlassian.net/browse/MAID-1363) remove wake_up.rs